Splitting reconciler and controller.
